+++ Assets/Scripts/MazeFlags.cs
@@ -17,14 +17,16 @@ public enum MazeFlags
 
 	PassagesDiagonal = 0b1111_0000,
 
-	VisibleToPlayer = 0b0001_0000_0000,
+	CutCorners = 0b1111_0000_0000,
 
-	VisbleToAgentA = 0b0010_0000_0000,
-	VisbleToAgentB = 0b0100_0000_0000,
-	VisbleToAgentC = 0b1000_0000_0000,
+	VisibleToPlayer = 0b0001_0000_0000_0000,
 
-	VisibleToAllAgents = 0b1110_0000_0000,
-	VisibleToAll = 0b1111_0000_0000
+	VisbleToAgentA = 0b0010_0000_0000_0000,
+	VisbleToAgentB = 0b0100_0000_0000_0000,
+	VisbleToAgentC = 0b1000_0000_0000_0000,
+
+	VisibleToAllAgents = 0b1110_0000_0000_0000,
+	VisibleToAll = 0b1111_0000_0000_0000
 }
 
 public static class MazeFlagsExtensions